Fixed XML output for perspectives
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@14215 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/io/layout/DockPanelData.java
@@ -161,13 +161,13 @@ public class DockPanelData {
 		sb.append("\" size=\"");
 		sb.append(getEmbeddedSize());
 		sb.append("\" window=\"");
-		sb.append(getFrameBounds().getX());
+		sb.append((int)getFrameBounds().getX());
 		sb.append(",");
-		sb.append(getFrameBounds().getY());
+		sb.append((int)getFrameBounds().getY());
 		sb.append(",");
-		sb.append(getFrameBounds().getWidth());
+		sb.append((int)getFrameBounds().getWidth());
 		sb.append(",");
-		sb.append(getFrameBounds().getHeight());
+		sb.append((int)getFrameBounds().getHeight());
 		sb.append("\" />\n");
 		return sb.toString();
 	}
